Changed Criminals BDI into Dustmen - stable
Criminals still work as in the version without BDI. The BDI extension on the criminals is now applied to the new class DustMan that collects thrash from Pubs. The console displays how much trash is accumulated in each Pub everytime it's open. The DustMen will collect all the trash, until all of it it's collected.