Stable version
ConcertHalls will now follow the music tastes of the PartyLovers. PartyLovers music tastes can vary through time.
ConcertHalls will discover the taste of the people through time with a simple RL algorithm that explores environment (discover of music tastes) and the chances of choosing the MusicGenre for the next cycle to exploit the knowledge acquired. To appreciate the solution, use greater value for nbPartyLovers.
A Perturbation is introduced to balance fashions exceeds.